pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sportscore"
version = "0.1.0"
description = "Shared core infrastructure for multi-sport analytics platforms"
requires-python = ">=3.12"
dependencies = [
    "pymongo>=4.0.0",
    "pandas>=1.3.0",
    "numpy>=1.20.0",
    "scikit-learn>=1.0.0",
    "pyyaml>=6.0",
    "pydantic>=1.10.0",
    "requests>=2.25.0",
    "pytz>=2023.3",
]

[project.optional-dependencies]
ml = [
    "xgboost>=1.5.0",
    "lightgbm>=3.3.0",
    "catboost>=1.0.0",
]
market = [
    "cryptography>=41.0.0",
]
all = [
    "sportscore[ml,market]",
]

[tool.setuptools.package-data]
sportscore = ["sports/*.yaml"]

[tool.setuptools.packages.find]
include = ["sportscore*"]
